# Override CopyTo to improve performance