If CopyTo is not overridden, it is used the default implementation in Stream class which is inefficient as it uses ArrayPool and writes in loop.