@lechu445 2023-02-17
before:
```
|    Method |     Mean |     Error |    StdDev | Allocated |
|---------- |---------:|----------:|----------:|----------:|
| WriteSpan | 1.508 us | 0.0459 us | 0.1272 us |     968 B |
```

after:
```
|    Method |     Mean |     Error |    StdDev |   Median | Allocated |
|---------- |---------:|----------:|----------:|---------:|----------:|
| WriteSpan | 1.042 us | 0.0266 us | 0.0715 us | 1.000 us |     632 B |
```

benchmark code:
```c#
[MemoryDiagnoser]
public class Test
{
    private RecyclableMemoryStream subject;
    private MemoryStream dest;
    private readonly byte[] bytes = new byte[] { 1, 2, 3, 4, 5, 6 };

    [IterationSetup]
    public void IterationSetup()
    {
        var manager = new RecyclableMemoryStreamManager();
        this.subject = new RecyclableMemoryStream(manager);
        for (int i = 0; i < 1000; i++)
        {
            this.subject.Write(bytes.AsSpan());
        }
        this.dest = new MemoryStream(10_000);
    }

    [Benchmark]
    public void WriteSpan() => this.subject.CopyTo(this.dest);
}
```