@lechu445 2023-02-17
after applying the suggestion, the performance is even better
```
|    Method |     Mean |    Error |   StdDev | Allocated |
|---------- |---------:|---------:|---------:|----------:|
| WriteSpan | 672.2 ns | 32.72 ns | 91.22 ns |     632 B | 
```